AlexNet Test: Retrying C=64, num_workers=8, batch_size=2048 (and n_epochs=20) but with a new version of knapsack_specialized that takes into account sparsity representation for big matrices

+++ utils/knapsack.py
@@ -536,8 +536,8 @@ def knapsack_specialized_pruning_sparse(xi, v, w, C, device, delta):
 
     # Fill the dense matrix from sparse representation
     x = torch.zeros((M, C), device=device, dtype=x_val.dtype)
-    x.scatter_add_(1, x_idx.unsqueeze(1), x_val.unsqueeze(1))
-    x.scatter_add_(1, x_idx_2.unsqueeze(1), x_val_2.unsqueeze(1))
+    x.scatter_add_(1, x_idx.unsqueeze(1).to(torch.long), x_val.unsqueeze(1))
+    x.scatter_add_(1, x_idx_2.unsqueeze(1).to(torch.long), x_val_2.unsqueeze(1))    
 
     return x, lambda_opt, objective_values
